tests

+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/repositories/ProductRepository.java
@@ -8,7 +8,7 @@ import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
 
     ArrayList<ProductEntity> findByName(String name);
